+++ comicbox/box/computed/date.py
@@ -1,8 +1,6 @@
 """Comicbox Computed Date tags."""
 
-from collections.abc import Callable
 from datetime import date
-from types import MappingProxyType
 from typing import Any
 
 from loguru import logger
@@ -15,7 +13,6 @@ from comicbox.formats.comicbox.schema import (
     MONTH_KEY,
     YEAR_KEY,
 )
-from comicbox.merge import AdditiveMerger, Merger
 
 _DATE_PART_KEYS = (YEAR_KEY, MONTH_KEY, DAY_KEY)
 
@@ -85,13 +82,3 @@ class ComicboxComputedDate(ComicboxComputedIssue):
         if not computed_date:
             return None
         return {DATE_KEY: computed_date}
-
-    COMPUTED_ACTIONS: MappingProxyType[str, tuple[Callable, type[Merger] | None]] = (
-        MappingProxyType(
-            {
-                # Order is important here
-                **ComicboxComputedIssue.COMPUTED_ACTIONS,
-                "from date": (_get_computed_from_date, AdditiveMerger),
-            }
-        )
-    )